Conditionally initialize database on first start
Before this update, db initialization was separate step accessed by
passing the --initialize flag to the run script. This still works.
But the initialize method has been made idempotent (probably) and
now runs before the server startup as well. This is helpful when we
want to use these containers with something like docker-compose,
where a separate initialization step isn't practical. Note that a
passphrase must be set and will be pulled from the PASSPHRASE env
variable. This must be passed in when using docker-compose.